Refactor: FileUploader - 파일 관련된 이벤트 훅으로 분리
