+++ src/components/FileUploader/index.tsx
@@ -1,9 +1,6 @@
-import { ONLY_UPLOAD_FILES_BELOW_10MB } from 'constants/errors';
-
 import { Typography } from '@material-ui/core';
 import { Image } from '@material-ui/icons';
-import { useCommonSnackbar } from 'components/CommonSnackbar/useCommonSnackbar';
-import { ChangeEvent, DragEvent, useRef, useState } from 'react';
+import { useRef } from 'react';
 import { theme } from 'themes';
 
 import {
@@ -12,18 +9,7 @@ import {
   UploadContainer,
   FileInput,
 } from './FileUploader.css';
-
-type FileMetadata = {
-  name: string;
-  size: number;
-  type: string;
-};
-
-export type FileInformation = {
-  url: string;
-  metadata: FileMetadata;
-  file: File;
-};
+import { FileInformation, useFileUploader } from './hooks/useFileUploader';
 
 type ImageAccept = 'png' | 'jpeg' | 'jpg' | 'gif';
 
@@ -49,76 +35,13 @@ const FileUploader = ({
   const hiddenFileInput = useRef<HTMLInputElement>(null);
   const types = imageAccepts.map((accept) => `image/${accept}`);
 
-  const [isOversized, setIsOversized] = useState(false);
-
-  useCommonSnackbar({
-    message: ONLY_UPLOAD_FILES_BELOW_10MB,
-    severity: 'error',
-    dependencies: [isOversized],
-    callbackAfterReset: () => setIsOversized(false),
+  const { onFileChange, onFileDrop } = useFileUploader({
+    maximumSize,
+    onChange,
   });
-
-  const validFileSize = (fileSize: number): boolean => {
-    return fileSize <= maximumSize;
-  };
-
-  const getFileMetadata = ({ name, size, type }: File) => {
-    return {
-      name,
-      size,
-      type,
-    };
-  };
-
-  const isInvalidFiles = (files: FileList | null): boolean => {
-    if (files == null || files.length < 1) {
-      return false;
-    }
-
-    const hasOversized = Array.from(files).some(
-      ({ size }) => !validFileSize(size),
-    );
-
-    if (hasOversized) {
-      setIsOversized(true);
-      return false;
-    }
-
-    return true;
-  };
-
-  const uploadFile = (files: FileList | null) => {
-    if (isInvalidFiles(files)) {
-      return;
-    }
-
-    const mappingFiles = Array.from(files ?? []).map(
-      (file): FileInformation => {
-        const metadata = getFileMetadata(file);
-        const url = URL.createObjectURL(file);
-
-        return { url, metadata, file };
-      },
-    );
-
-    onChange(mappingFiles);
-  };
-
-  const handleFileChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const { files } = event.target ?? {};
-
-    uploadFile(files);
-  };
 
   const handleOnDropZoneClick = (): void => {
     hiddenFileInput?.current?.click();
-  };
-
-  const handleOnDrop = (event: DragEvent<HTMLDivElement>): void => {
-    event.preventDefault();
-    const files = event.dataTransfer.files;
-
-    uploadFile(files);
   };
 
   return (
@@ -138,7 +61,7 @@ const FileUploader = ({
       })}
       <DropZone
         onClick={handleOnDropZoneClick}
-        onDrop={handleOnDrop}
+        onDrop={onFileDrop}
         onDragOver={(event) => event.preventDefault()}
         width={width}
         height={height}
@@ -154,7 +77,7 @@ const FileUploader = ({
           type="file"
           accept={types.join(',')}
           ref={hiddenFileInput}
-          onChange={handleFileChange}
+          onChange={onFileChange}
           multiple={isMultiple}
         />
       </DropZone>
